move draw_center_panel from app.rs into new file app_center.rs
